project configuration

+++ src/interfaces/Asset.java
@@ -8,15 +8,18 @@ package interfaces;
  * its price, owner and user can be modified and got.
  *
 */
+
 import java.io.Serializable;
 
 public interface Asset extends Serializable {
     //return an array to store the information of the asset,arr[0] is the rent price, arr[1] is the sell price,
     // arr[2] is the cost price,
     double[] getPrice();
-    void setPrice(double rentPrice,double sellPrice,double costPrice);
+    void setPrice(int rentPrice,int sellPrice,int costPrice);
     Person getOwner();
     void setOwner(Person owner);
     Person getUser();
     void setUser(Person user);
+
+
 }
